refactor: move some packages
